Add ptrac export
Add option to save each report parsed as a ptrac file in addition to importing data into a PT instance.

After the CSV is parsed you will be given 2 prompts to import data to PT and to save data to ptrac(s). Both prompts are asked regardless of each other's choice.