Update readme and examples - Start tracking api folder
Updated the README.md to describe this reports

The api folder is generated from PT's Postman Collection. This stores the URLs for each endpoint and acts as an API wrapper library. The latest version of this folder will be tracked in this repo. This will be updated when changes are made to the Postman Collection. The Postman Collection is the official API endpoint docs for PT.