update api folder
The Postman folder structure was recently overhauled to organize endpoints by category. This was done since the distinction between v1 and v2 endpoints didn't denote different versions. Updates the collection of API endpoints in the api folder to match the `20230815 categories` update.

Modify the auth_utils.py file to implement the new endpoint function calls to authenticate a user